Add patched curlconverter
Since `web-tree-sitter` from NPM is not a ES module, vite cannot extract
`tree-sitter.wasm` into an asset, we use `exec` protocol to build an
all-in-one version for ourselves.

Moreover, we also build `curlconverter` without top-level-await for
ourselves because rollup cannot generate code in IIFE format if there is
a top-level-await statement.
And, we need vite to extract `tree-sitter-bash.wasm` into an asset.